totalne predelan skills system, proprietarni skilly jsou zachovano, ale pribyli konfigurovatelne skily v xml
diky konfigurovatelnym skilum se nabizi neskutecne mnoho moznosti skilu,
navic kazdy konfigurovatelny skill si muze sam definvat kolikrat se
pouzije nebo jak moc silny muze byt. dale take predelana sila schopnosti
podle based atrobitu ktere se deli na skill_power, attack a skill, je
treba vymyslem metematicke rovnice na vypocet sily schopnosti podle
techto atributu tak aby to bylo balanced

+++ src/com/nex/gamebook/game/EnemyAssign.java
@@ -3,9 +3,9 @@ package com.nex.gamebook.game;
 public class EnemyAssign {
 
 	private String enemyKey;
-	private double xpcoeff = 0f;
+	private float xpcoeff = 0f;
 
-	public EnemyAssign(String enemyKey, double xpcoeff) {
+	public EnemyAssign(String enemyKey, float xpcoeff) {
 		super();
 		this.enemyKey = enemyKey;
 		this.xpcoeff = xpcoeff;
